Add files via upload
This is my fun take on a basic hello world python script with a way for me to understand how naming the variable the same thing post "print" action, leads to a different output and overwrites the existing information but not always, just when it's set up that way. It was my first Python script and I had fun with it.